css/css-fonts/parsing/font-family-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-family with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-family-prop">
<meta name="assert" content="font-family supports the full grammar '[ <family-name> | <generic-family> ] #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("font-family", "Helvetica");
test_valid_value("font-family", "Helvetica, Verdana");

test_valid_value("font-family", "serif");
test_valid_value("font-family", "serif, sans-serif, cursive, fantasy, monospace");

test_valid_value("font-family", "Verdana, sans-serif, Verdana, sans-serif");
</script>
</body>
</html>
